refactor: header 컴포넌트를 각 페이지별로 이동
