+++ web/src/components/auth/AuthHeader.module.css
@@ -1,0 +1,11 @@
+.headerName {
+  position: absolute;
+  transform: translate(-50%, -50%);
+  top: 50%;
+  left: 50%;
+
+  font-weight: 600;
+  font-size: 20px;
+  line-height: 24px;
+  color: #1c1b1f;
+}
