+++ web/src/components/common/header/Header.module.css
@@ -1,0 +1,8 @@
+.wrapper {
+  position: relative;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  height: 70px;
+  padding: 0 20px;
+}
