--- web/src/components/header/Header.module.css
@@ -1,34 +1,0 @@
-.wrapper {
-  position: relative;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  height: 70px;
-  padding: 0 20px;
-}
-
-.h1 {
-  margin-left: 10px;
-  font-weight: 400;
-  font-size: 18px;
-  line-height: 28px;
-  text-align: center;
-  color: #1c1b1f;
-}
-
-.right {
-  display: flex;
-  align-items: center;
-}
-
-.headerName {
-  position: absolute;
-  transform: translate(-50%, -50%);
-  top: 50%;
-  left: 50%;
-
-  font-weight: 600;
-  font-size: 20px;
-  line-height: 24px;
-  color: #1c1b1f;
-}
