+++ web/src/components/main/MainHeader.module.css
@@ -1,0 +1,13 @@
+.h1 {
+  margin-left: 10px;
+  font-weight: 400;
+  font-size: 18px;
+  line-height: 28px;
+  text-align: center;
+  color: #1c1b1f;
+}
+
+.right {
+  display: flex;
+  align-items: center;
+}
