+++ web/src/containers/Main/index.tsx
@@ -1,5 +1,5 @@
-import Header from '@/components/header/Header';
-import Card from '@/components/card/Card';
+import Header from '@/components/main/Header';
+import Card from '@/components/common/card/Card';
 import styles from '@/styles/Main.module.css';
 import { useSelector } from 'react-redux';
 import { userState } from '@/modules/user';
@@ -12,6 +12,7 @@ export default function MainContainer() {
   }));
   return (
     <main>
+      <Header />
       <section className={styles.bannerWrapper}></section>
       <section className={styles.cardWrapper}>
         <Card card={'asdf'} />
